+++ src/components/IconStyles.tsx
@@ -1,0 +1,15 @@
+import { StyleSheet } from 'react-native';
+import { MATERIAL_ICON_SIZE } from '../constants/dimensions';
+
+export const styles = StyleSheet.create({
+  searchIcon: {
+    width: MATERIAL_ICON_SIZE
+  },
+  backIcon: {
+    width: MATERIAL_ICON_SIZE
+  },
+  clearFilterIcon: {
+    width: MATERIAL_ICON_SIZE,
+    marginRight: 10
+  }
+});
